+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 * BREAKING: Team Members werden mit Name gespeichert -> Anpassung Teams Schema
 * Info in der Reception, wenn mit Bildschirmen < 1024 Pixel Auflösung gespielt werden will
 * Darstellung der hochgeladenen Bilder verbessert, mit Viewer
+  * Bilder können ausgeblendet werden (sind dann für andere Teams nicht sichtbar)
+
 
 ## v3.5.2 9.6.25
 * Bugfix in Marketplace (Variable undefined)